schedule_print.cc: print each map of an extension on a new line
This was omitted in the original commit in the prehistory that made
union_* members of schedule nodes print members of a union on separate
lines.  Since there is no isl_union_map_get_map_list, wrap the union_map
to obtain a union_set, get the list of sets and unwrap them to get the
maps back.